% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmrlandings-package.R
\docType{package}
\name{dmrlandings}
\alias{dmrlandings-package}
\alias{dmrlandings}
\title{R tools to ease mining DMR landings data.}
\description{
R tools to ease mining DMR landings data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigelowlab/dmrlandings}
}

}
