% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr.R
\name{annualize_dmr_landings}
\alias{annualize_dmr_landings}
\title{Aggregate data to state-wide annual totals per species}
\usage{
annualize_dmr_landings(x = read_dmr_landings(when = "merged"))
}
\arguments{
\item{x}{data frame of dmr landings (merged)}
}
\value{
data frame
}
\description{
Aggregate data to state-wide annual totals per species
}
