% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr.R
\name{read_dmr_landings}
\alias{read_dmr_landings}
\title{Read ME DMR landings data}
\usage{
read_dmr_landings(when = c("modern", "historic", "merged", "annualized")[4])
}
\arguments{
\item{when}{chr, one of "modern" (default), "historic", "merged", "annualized"}
}
\value{
data frame
}
\description{
\href{https://mainedmr.shinyapps.io/Landings_Portal/}{Maine's DMR landings data portal} allows
for the download of "historic" (state wide annual totals per species), "modern"
(per port per species annual totals) or "annualized" (statewide totals per year per species)
landings data.  Use this function to read historic, modern, merged or annualized data.
}
